[feat]: impl deleted recipes

+++ composeApp/src/commonMain/kotlin/ru/topbun/cherry_tip/presentation/ui/components/Recipe.kt
@@ -237,7 +237,7 @@ fun RowScope.NutrientsItem(
 @Composable
 fun RowScope.NutrientsItem(
     text: String,
-    value: Int,
+    value: Float,
     progress: Float,
     progressColor: Color
 ) {
